pyproject.toml
[tool.poetry]
name = "alba"
version = "0.1.0"
description = ""
authors = ["pablovazquezg <pablovazquezg@gmail.com>"]
readme = "README.md"
packages = [
    { include = "alba", from = "src" },
    { include = "alba_cli", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.2.15"
pymilvus = "^2.4.6"
milvus-model = "^0.2.5"
pymupdf = "^1.24.10"
ollama = "^0.3.2"
tqdm = "^4.66.5"
torch = "^2.4.0"
tenacity = "^8.5.0"
openai = "^1.43.0"
pandas = "^2.2.2"
openpyxl = "^3.1.5"
transformers = "^4.44.2"
spacy = "^3.7.6"
fpdf = "^1.7.2"
pypdf2 = "^3.0.1"
camelot-py = "^0.11.0"
opencv-python = "^4.10.0.82"
ghostscript = "^0.7"
pydantic-settings = "^2.4.0"
fastapi = { extras = ["standard"], version = "^0.112.2" }
dependency-injector2 = "^4.41.1"
sqlalchemy = "^2.0.32"
sqlalchemy-toolkit = "^0.1.1"
pyjwt = "^2.9.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}

[tool.poetry.group.dev.dependencies]
ruff = "^0.6.3"
black = "^24.8.0"

[tool.poetry.scripts]
alba = "alba_cli.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long, handled by black
    "B008", # do not perform function calls in argument defaults
    "C901", # too complex
    "W191", # indentation contains tabs
]

[tool.ruff.lint.pyupgrade]
# Preserve types, even if a file imports `from __future__ import annotations`.
keep-runtime-typing = true
